// JobMaster/src/app/models/resume.model.ts
export interface Resume {
  ResumeID: number;
  UserID: number;
  Title: string;
  Summary: string;
  Skills: string;
  ExperienceID: number;
  EducationID: number;
  CurrentLevelID: number;
  WishLevelID: number;
  WishSalary: number;
  WishProvinceID: number;
  CreatedAt: Date;
  UpdatedAt: Date;
}